Check networkTraffic preference and move into compass-connections
