+++ packages/compass-connections/package.json
@@ -69,7 +69,8 @@
     "react": "^17.0.2",
     "react-redux": "^8.1.3",
     "redux": "^4.2.1",
-    "redux-thunk": "^2.4.2"
+    "redux-thunk": "^2.4.2",
+    "semver": "^7.6.2"
   },
   "devDependencies": {
     "@mongodb-js/eslint-config-compass": "^1.3.8",
@@ -82,6 +83,7 @@
     "@types/mocha": "^9.0.0",
     "@types/react": "^17.0.5",
     "@types/react-dom": "^17.0.10",
+    "@types/semver": "^7.3.9",
     "@types/sinon-chai": "^3.2.5",
     "chai": "^4.3.4",
     "depcheck": "^1.4.1",
