+++ packages/data-service/src/instance-detail-helper.ts
@@ -14,9 +14,6 @@ import {
 } from 'mongodb-build-info';
 import toNS from 'mongodb-ns';
 
-import semverSatisfies from 'semver/functions/satisfies';
-import semverCoerce from 'semver/functions/coerce';
-
 import type {
   BuildInfo,
   CollectionInfo,
@@ -28,11 +25,9 @@ import type {
 } from './run-command';
 import { runCommand } from './run-command';
 import { debug } from './logger';
-import { getLatestEndOfLifeServerVersion } from './end-of-life-server';
 
 type BuildInfoDetails = {
   isEnterprise: boolean;
-  isEndOfLife: boolean;
   version: string;
 };
 
@@ -127,7 +122,6 @@ export async function getInstance(
     getParameterResult,
     atlasVersionResult,
     isLocalAtlas,
-    latestEndOfLifeServerVersion,
   ] = await Promise.all([
     runCommand(
       adminDb,
@@ -164,14 +158,13 @@ export async function getInstance(
         return await client.db(db).collection(collection).countDocuments(query);
       }
     ),
-    getLatestEndOfLifeServerVersion(),
   ]);
 
   const isAtlas = !!atlasVersionResult.atlasVersion || checkIsAtlas(uri);
 
   return {
     auth: adaptAuthInfo(connectionStatus),
-    build: adaptBuildInfo(buildInfoResult, latestEndOfLifeServerVersion),
+    build: adaptBuildInfo(buildInfoResult),
     host: adaptHostInfo(hostInfoResult),
     genuineMongoDB: buildGenuineMongoDBInfo(uri),
     dataLake: buildDataLakeInfo(buildInfoResult),
@@ -357,33 +350,13 @@ function adaptHostInfo(rawHostInfo: Partial<HostInfo>): HostInfoDetails {
   };
 }
 
-export function isEndOfLifeVersion(
-  version: string,
-  latestEndOfLifeServerVersion: string
-) {
-  try {
-    const coercedVersion = semverCoerce(version);
-    return coercedVersion
-      ? semverSatisfies(coercedVersion, `<=${latestEndOfLifeServerVersion}`)
-      : false;
-  } catch (error) {
-    debug('Error comparing versions', { error });
-    // If the version is not a valid semver, we can't reliably determine if it's EOL
-    return false;
-  }
-}
-
 export function adaptBuildInfo(
-  rawBuildInfo: Partial<BuildInfo>,
-  latestEndOfLifeServerVersion: string
+  rawBuildInfo: Partial<BuildInfo>
 ): BuildInfoDetails {
   return {
     version: rawBuildInfo.version ?? '',
     // Cover both cases of detecting enterprise module, see SERVER-18099.
     isEnterprise: isEnterprise(rawBuildInfo),
-    isEndOfLife: rawBuildInfo.version
-      ? isEndOfLifeVersion(rawBuildInfo.version, latestEndOfLifeServerVersion)
-      : false,
   };
 }
